refactor(rendering): Update the startup rendering
