+++ analysis/bundle-size-analyzer.log
@@ -1,26 +1,26 @@
 leaflet: 433.27 KB (15.5%)
 proj4: 337.93 KB (12.1%)
 @material-ui/core: 312.64 KB (11.2%)
-esri-leaflet: 272.49 KB (9.74%)
+esri-leaflet: 272.49 KB (9.73%)
 esri-leaflet-renderers: 146.45 KB (5.23%)
-@material-ui/styles: 144.3 KB (5.16%)
+@material-ui/styles: 144.3 KB (5.15%)
 react-dom: 119.19 KB (4.26%)
-popper.js: 86.41 KB (3.09%)
+popper.js: 86.41 KB (3.08%)
 i18next: 73.4 KB (2.62%)
 jss: 56.04 KB (2.00%)
 react-transition-group: 54.62 KB (1.95%)
 react-i18next: 27.86 KB (0.995%)
-mgrs: 20.46 KB (0.731%)
-react-leaflet: 20.27 KB (0.724%)
+mgrs: 20.46 KB (0.730%)
+react-leaflet: 20.27 KB (0.723%)
 @material-ui/system: 18.14 KB (0.648%)
-@babel/runtime: 14.69 KB (0.525%)
-css-vendor: 14.49 KB (0.518%)
+@babel/runtime: 14.69 KB (0.524%)
+css-vendor: 14.49 KB (0.517%)
 i18next-http-backend: 11.82 KB (0.422%)
 @react-leaflet/core: 11.6 KB (0.414%)
-wkt-parser: 11.41 KB (0.408%)
+wkt-parser: 11.41 KB (0.407%)
 @terraformer/arcgis: 11.37 KB (0.406%)
 @material-ui/utils: 9.65 KB (0.345%)
-react: 7.63 KB (0.273%)
+react: 7.63 KB (0.272%)
 proj4leaflet: 7.56 KB (0.270%)
 style-loader: 6.67 KB (0.238%)
 @material-ui/icons: 6.29 KB (0.225%)
@@ -31,14 +31,14 @@ jss-plugin-global: 4.01 KB (0.143%)
 screenfull: 3.78 KB (0.135%)
 jss-plugin-nested: 3.55 KB (0.127%)
 react-is: 2.69 KB (0.0959%)
-hoist-non-react-statics: 2.68 KB (0.0957%)
-prop-types: 2.58 KB (0.0923%)
-css-loader: 2.38 KB (0.0850%)
+hoist-non-react-statics: 2.68 KB (0.0956%)
+prop-types: 2.58 KB (0.0922%)
+css-loader: 2.38 KB (0.0849%)
 jss-plugin-rule-value-function: 2.3 KB (0.0820%)
 object-assign: 2.06 KB (0.0735%)
 tiny-binary-search: 1.93 KB (0.0691%)
-jss-plugin-vendor-prefixer: 1.49 KB (0.0531%)
-jss-plugin-camel-case: 1.46 KB (0.0521%)
+jss-plugin-vendor-prefixer: 1.49 KB (0.0530%)
+jss-plugin-camel-case: 1.46 KB (0.0520%)
 dom-helpers: 1.21 KB (0.0432%)
 node-fetch: 732 B (0.0255%)
 clsx: 659 B (0.0230%)
@@ -47,4 +47,4 @@ is-in-browser: 510 B (0.0178%)
 hyphenate-style-name: 452 B (0.0158%)
 void-elements: 376 B (0.0131%)
 tiny-warning: 361 B (0.0126%)
-<self>: 510.72 KB (18.2%)
+<self>: 512.68 KB (18.3%)
