+++ public/index.html
@@ -1,40 +1,59 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title><%= htmlWebpackPlugin.options.title %></title>
-    <link rel="stylesheet" href="/css/style.css" />
-    <link
-      href="https://fonts.googleapis.com/css?family=Roboto|Montserrat:200,300,400,900|Merriweather"
-      rel="stylesheet"
-      type="text/css"
-    />
-    <style>
-        #mapid { height: 1000px; }
-    </style>
-  </head>
-  <body>
-    <div id="root"></div>
-    <span>Leaflet samples with VueJS framework (English map in Web Mercator)</span>
-    <div id="mapWM" class="llwp-map" data-leaflet="{ 'name': 'Web Mercator', 'projection': 3857, 'zoom': 4, 'center': [60,-100], 'language': 'en-CA',
+    <head>
+        <meta charset="UTF-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <title><%= htmlWebpackPlugin.options.title %></title>
+        <link rel="stylesheet" href="/css/style.css" />
+        <link
+            href="https://fonts.googleapis.com/css?family=Roboto|Montserrat:200,300,400,900|Merriweather"
+            rel="stylesheet"
+            type="text/css"
+        />
+        <style>
+            #mapid {
+                height: 1000px;
+            }
+        </style>
+    </head>
+    <body>
+        <div id="root"></div>
+        <div>
+            <h3>The Atlas of Canada - Clean Energy Resources and Projects (CERP)</h3>
+            <div
+                id="mapWM"
+                class="llwp-map"
+                data-leaflet="{ 'name': 'Web Mercator', 'projection': 3857, 'zoom': 4, 'center': [60,-100], 'language': 'en-CA',
         'layers':[
             { 'url': 'https://webservices.maps.canada.ca/arcgis/services/StatCan/census_subdivisions_2016_en/MapServer/WMSServer', 'type': 'ogcWMS', 'entries': '0' },
             { 'url': 'https://geoappext.nrcan.gc.ca/arcgis/rest/services/NRCAN/Investing_Energy_Canada_en/MapServer', 'type': 'esriDynamic', 'entries': '0, 2' },
             { 'url': 'https://geoappext.nrcan.gc.ca/arcgis/rest/services/GSCC/Geochronology/MapServer', 'type': 'esriDynamic', 'entries': '0' },
             { 'url': 'https://geo.weather.gc.ca/geomet', 'type': 'ogcWMS', 'entries': 'RAQDPS-FW.CE_PM2.5-DIFF-YAvg' }
-        ]}">
-    </div>
-    <span>Samples to show a second map on the same App with different projection (French map in LCC)</span>
-    <div id="mapLCC" class="llwp-map" data-leaflet="{ 'name': 'LCC', 'projection': 3978, 'zoom': 12, 'center': [45,-75], 'language': 'fr-CA',
+        ]}"
+            ></div>
+            <p>
+                This map is an example of a Leaflet map done with React. This map is in Web Mercator projection with an English interface.
+            </p>
+        </div>
+        <hr />
+        <div>
+            <h3>L'Atlas du Canada - Ressources et projets d'énergie propre (RPEP)</h3>
+            <div
+                id="mapLCC"
+                class="llwp-map"
+                data-leaflet="{ 'name': 'LCC', 'projection': 3978, 'zoom': 12, 'center': [45,-75], 'language': 'fr-CA',
         'layers':[
             { 'url': 'https://webservices.maps.canada.ca/arcgis/services/StatCan/census_subdivisions_2016_en/MapServer/WMSServer', 'type': 'ogcWMS', 'entries': '0' },
             { 'url': 'https://geoappext.nrcan.gc.ca/arcgis/services/MMS/T100EP_E/MapServer/WMSServer', 'type': 'ogcWMS', 'entries': '0' },
             { 'url': 'https://geoappext.nrcan.gc.ca/arcgis/rest/services/NACEI/energy_infrastructure_of_north_america_en/MapServer/1', 'type': 'esriFeature' },
             { 'url': 'https://geoappext.nrcan.gc.ca/arcgis/rest/services/NACEI/energy_infrastructure_of_north_america_en/MapServer', 'type': 'esriDynamic', 'entries': '4, 5, 7, 8, 9, 10, 12, 13, 14, 15, 16, 28, 29, 37' },
             { 'url': 'https://geoappext.nrcan.gc.ca/arcgis/rest/services/GSCC/Geochronology/MapServer', 'type': 'esriDynamic', 'entries': '0' }
-        ]}">
-    </div>
-    <div id="vanilla"></div>
-  </body>
+        ]}"
+            ></div>
+            <p>
+                Cette carte est un exemple d'une carte créée à partir de Leaflet et React. La projection LCC est utilisée et l'interface est
+                en français.
+            </p>
+        </div>
+    </body>
 </html>
